traversal: pass node retrieval error to walker functions
Currently traversing nodes any error encountered while loading a link or decoding a node
will result in an error being returned within the walk function which causes the walk to
exit. This makes it difficult to process a heterogenous chain of nodes that include types
not supported by available codecs.

This change extends the existing walk callback functions WalkFn, TransformFn and AdvVisitFn
by adding an error argument that will contain any error encountered while attempting to
visit the node. This allows the function to decide how to handle the error.

The Go standard library follows this pattern with the WalkFunc type in path/filepath.